+++ central/node/datastore/dackbox/datastore/datastore_impl_test.go
@@ -28,6 +28,7 @@ import (
 	"github.com/stackrox/rox/pkg/dackbox/edges"
 	"github.com/stackrox/rox/pkg/dackbox/indexer"
 	"github.com/stackrox/rox/pkg/dackbox/utils/queue"
+	"github.com/stackrox/rox/pkg/features"
 	"github.com/stackrox/rox/pkg/rocksdb"
 	"github.com/stackrox/rox/pkg/sac"
 	"github.com/stackrox/rox/pkg/scancomponent"
@@ -53,6 +54,11 @@ type NodeDataStoreTestSuite struct {
 }
 
 func (suite *NodeDataStoreTestSuite) SetupSuite() {
+	if features.PostgresDatastore.Enabled() {
+		suite.T().Skip("Skip dackbox tests if postgres is enabled")
+		suite.T().SkipNow()
+	}
+
 	suite.db = rocksdbtest.RocksDBForT(suite.T())
 
 	suite.indexQ = queue.NewWaitableQueue()